fix: Room price 타입 변경 long -> BigDecimal

+++ src/main/java/com/ryan/roomreservationservice/domain/Accommodation.java
@@ -4,6 +4,7 @@ import com.ryan.roomreservationservice.domain.enums.AccommodationStatus;
 import com.ryan.roomreservationservice.utils.exception.ErrorMessage;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -42,11 +43,11 @@ public class Accommodation {
         accommodation.status = AccommodationStatus.COMPLETED;
     }
 
-    public long getPaymentAmount(LocalDateRange reservationDate) {
+    public BigDecimal getPaymentAmount(LocalDateRange reservationDate) {
         return this.room.calculateRoomPaymentAmount(reservationDate);
     }
 
-    public long getRefundAmount(LocalDate cancelLocalDate, LocalDateRange reservationDate) {
+    public BigDecimal getRefundAmount(LocalDate cancelLocalDate, LocalDateRange reservationDate) {
         return this.room.calculateRoomRefundAmount(cancelLocalDate, reservationDate);
     }
 }
